requirements.txt
# Core dependencies
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
yfinance>=0.2.18
requests>=2.28.2
python-dotenv>=1.0.0
PyYAML>=6.0

# Machine Learning
scikit-learn>=1.2.2
torch>=2.0.1
transformers>=4.30.2

# LLM and NLP
openai>=0.27.8
nltk>=3.8.1
spacy>=3.5.3

# Data Processing
beautifulsoup4>=4.12.2
newsapi-python>=0.2.7
alpha_vantage>=2.3.1

# Visualization
plotly>=5.14.1
matplotlib>=3.7.1
seaborn>=0.12.2

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0 